requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Advanced ML
transformers>=4.30.0
diffusers>=0.18.0
accelerate>=0.20.0
datasets>=2.12.0

# LLM & NLP
openai>=1.0.0
tiktoken>=0.4.0
langchain>=0.0.200
sentence-transformers>=2.2.0

# Scientific Computing
scipy>=1.10.0
networkx>=3.1.0
manifold-learn>=1.3.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
hydra-core>=1.3.0
rich>=13.4.0
typer>=0.9.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0
jupyterlab>=4.0.0
